added buyer activity
